+++ src/modules/bill_item/bill_item.model.js
@@ -5,8 +5,8 @@ const BillItem = db.define(
   'BillItem',
   {
     count: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 0 },
-    values: { type: DataTypes.INTEGER, allowNull: true, defaultValue: 0 },
-    value: { type: DataTypes.INTEGER, allowNull: true, defaultValue: 0 },
+    values: { type: DataTypes.BIGINT, allowNull: true, defaultValue: 0 },
+    value: { type: DataTypes.BIGINT, allowNull: true, defaultValue: 0 },
     note: { type: DataTypes.STRING, allowNull: true },
     colors: { type: DataTypes.STRING, allowNull: true },
     sizes: { type: DataTypes.STRING, allowNull: true },
