+++ src/modules/myBalance/MyBalance.model.js
@@ -4,8 +4,8 @@ const db = require('../../config/db');
 const MyBalance = db.define(
   'MyBalance',
   {
-    todayValues: { type: DataTypes.INTEGER, allowNull: true, defaultValue: 0 },
-    todayValue: { type: DataTypes.INTEGER, allowNull: true, defaultValue: 0 },
+    todayValues: { type: DataTypes.BIGINT, allowNull: true, defaultValue: 0 },
+    todayValue: { type: DataTypes.BIGINT, allowNull: true, defaultValue: 0 },
     yesterdayValues: {
       type: DataTypes.BIGINT,
       allowNull: true,
